seggregating data
